requirements.txt
python-telegram-bot==22.0
requests==2.31.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.2.0
schedule==1.2.0
vercel-kv==0.1.1 